// src/app/shared/models/transports/exports/dispatch-order-fee-export.model.ts
import { StringDecoder } from "string_decoder";

export interface DispatchOrderFeeExport {
  totalRows?: number;
  idCv?: number;
  refNo?: string;
  jobId?: string;
  referCode?: string;
  workflowId?:number;
  shippingTaskId?:number;
  makh?:string;
  tenkh?:string;
  loaixe?:string;
  bks?:string;
  laixe?:string;
  giadau?:number;
  contType?:string;
  loaiCont?:string;
  km?:number;
  tongKm?:number;
  dinhmucdau?:number;
  dieuchinh?:number;
  nhathauphu?: string;
  xethue?: boolean;
  pallets?: number;
  containerNumber?:string;
  sealNumber?:string;
  pickupTimes?:string;
  contSeals?:string;
  containers?:number;
  kmdau?:number;
  kmcuoi?:number;
  cungduongtoantuyen?: string;
  tomtatlenh?: string;
  cungduongtinhphi?: string;
  tollRouteCode?: string;
  ngaytao?: string;
  nguoitao?: string;
  chinhanh?: string;
  maphi?: string;
  tenmaphi?: string;
  soluong?: number;
  tien?: number;
  vat?: number;
  tongtien?: number;
  ghichu?: string;
  routeCode?:string;
  routeName?:string;
  inquiryTimeToTheFactory?:string;
  loaiCungduong?:string;
  noigiaonhan?:string;
  tronve?:boolean;
  luonghang?: string;
  loaihinh?: string;
  luonghangLuotdiName?: string;
  dinhmucDauLuotDi?: number;
  chang1LuotdiName?: string;
  chang1KmLuotDi?: number;
  chang2LuotdiName?: string;
  chang2KmLuotDi?: number;
  dieuchinhKmLuotDi?: number;
  luonghangTrungchuyenCangName?: string;
  dinhmucDauTrungchuyenCang?: number;
  kmTrungchuyenCang?: number;
  cang1LuotdiName?: string;
  cang2LuotdiName?: string;
  luonghangTrungchuyenNhamayName?: string;
  dinhmucDauTrungchuyenNhamay?: number;
  nhamay1LuotdiName?: string;
  nhamay2LuotdiName?: string;
  kmTrungchuyenNhamay?: number;
  luonghangLuotveName?: string;
  dinhmucDauLuotVe?: number;
  chang1LuotveName?: string;
  chang1KmLuotVe?: number;
  chang2LuotveName?: string;
  chang2KmLuotVe?: number;
  luonghangTrungchuyenCangveName?: string;
  dinhmucDauTrungchuyenCangVe?: number;
  cang1LuotveName?: string;
  cang2LuotveName?: string;
  kmTrungchuyenCangVe?: number;
  luonghangTrungchuyenNhamayveName?: string;
  dinhmucDauTrungchuyenNhamayVe?: number;
  nhamay1LuotveName?: string;
  nhamay2LuotveName?: string;
  kmTrungchuyenNhamayVe?: number;
  rStatus?: string;
}